test templates moved into own  package
